🔧 chore(test): relocate testing functions to dedicated script
- create new script test_tool_call_interceptor.py for testing functions
- remove test cases and related code from output_handle.py for better organization
